+++ lib/modules/declare_info_630b/ui/declare_info_group_ext_630b.dart
@@ -420,7 +420,7 @@ extension DeclareInfoGruopExt630b on DeclareInfo630bPage {
           return LocaleKeys.declareInfo_countDayInvalid.tr;
         }
         // REF: BHW-3106
-        final newText = double.tryParse(trimmedValue);
+        final newText = int.tryParse(trimmedValue);
         if (newText != null && newText == 0) {
           return LocaleKeys.declareInfo_countDayInvalid.tr;
         }
