+++ lib/modules/declare_info_630c/ui/declare_info_group_ext_630c.dart
@@ -288,7 +288,7 @@ extension DeclareInfoGroupExt630c on DeclareInfo630cPage {
           return LocaleKeys.declareInfo_countDayInvalid.tr;
         }
         // REF: BHW-3106
-        final newText = double.tryParse(trimmedValue);
+        final newText = int.tryParse(trimmedValue);
         if (newText != null && newText == 0) {
           return LocaleKeys.declareInfo_countDayInvalid.tr;
         }
